refactor: move AnchorTarget into SharedComponents
