+++ src/components/PortableText/PortableText.tsx
@@ -15,10 +15,10 @@ import { CTAInternalLinkEntry } from "../../common-lib/cms-types";
 import { LI, OL, P, Span } from "../Typography";
 import OakLink from "../OakLink";
 import getProxiedSanityAssetUrl from "../../common-lib/urls/getProxiedSanityAssetUrl";
-import AnchorTarget from "../AnchorTarget";
 
 import { PTActionTrigger } from "./PTActionTrigger";
 
+import AnchorTarget from "@/components/SharedComponents/AnchorTarget";
 import Box from "@/components/SharedComponents/Box";
 
 const reportError = errorReporter("PortableText");
