+++ src/components/Posts/PostPortableText/PostFootnotes.tsx
@@ -6,8 +6,7 @@ import type {
   PortableTextMarkDefinition,
 } from "@portabletext/types";
 
-import AnchorTarget from "../../AnchorTarget";
-
+import AnchorTarget from "@/components/SharedComponents/AnchorTarget";
 import Box from "@/components/SharedComponents/Box";
 
 export type Footnote = {
