+++ src/components/pages/LandingPages/SignUpForm.tsx
@@ -2,9 +2,9 @@ import { FC, useId } from "react";
 
 import Card from "../../Card";
 import NewsletterForm, { useNewsletterForm } from "../../Forms/NewsletterForm";
-import AnchorTarget from "../../AnchorTarget";
 import CardTitle from "../../Card/CardComponents/CardTitle";
 
+import AnchorTarget from "@/components/SharedComponents/AnchorTarget";
 import Box from "@/components/SharedComponents/Box";
 
 export const SignUpForm: FC<{ formTitle: string }> = ({ formTitle }) => {
